pyproject.toml
[project]
name = "ai-engine"
version = "0.1.0"
description = "AI Engine using LangChain + ChatOpenAI + LangServe"
authors = [
    { name = "Zhang", email = "zh_ang@aliyun.com" }
]
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "langchain (>=1.2.12,<2.0.0)",
    "openai (>=2.28.0,<3.0.0)",
    "langserve (>=0.3.3,<0.4.0)",
    "langchain-openai (>=1.1.11,<2.0.0)",
    "sse-starlette (>=3.3.2,<4.0.0)",
    "chromadb (>=1.5.5,<2.0.0)",
    "langchain-chroma (>=1.1.0,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "pypdf (>=6.9.1,<7.0.0)",
    "unstructured (>=0.21.5,<0.22.0)",
    "dashscope (>=1.25.14,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"